Refactor dark mode bits into a proper split out module.
